refactor(View): InputView와 OutputView 로직 수정
