+++ src/main/java/lotto/Application.java
@@ -1,20 +1,16 @@
 package lotto;
 
-import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 import lotto.service.LottoService;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
-        int purchaseAmount = Integer.parseInt(InputView.readPurchaseAmount());
+        int purchaseAmount = InputView.readPurchaseAmount();
 
         LottoNumberGenerator lottoNumberGenerator = new LottoNumberGenerator();
         LottoService lottoService = new LottoService(lottoNumberGenerator);
-
         List<Lotto> lottos = lottoService.purchaseLottos(purchaseAmount);
 
         OutputView.printLottoCount(lottos.size());
@@ -22,29 +18,15 @@ public class Application {
             OutputView.printLotto(lotto);
         }
 
-        Lotto winningNumbers = new Lotto(toInteger(InputView.readWinningNumbers()));
-        Integer bonusNumber = Integer.parseInt(InputView.readBonusNumber());
+        Lotto winningNumbers = new Lotto(InputView.readWinningNumbers());
+        Integer bonusNumber = InputView.readBonusNumber();
         WinningLotto winningLotto = new WinningLotto(winningNumbers, bonusNumber);
 
         LottoResult lottoResult = lottoService.calculateResults(lottos, winningLotto);
         printStatistics(lottoResult, purchaseAmount);
     }
 
-    public static List<Integer> toInteger(String winningNumbers) {
-        List<String> splittedNumbers = Arrays.asList(winningNumbers.split(","));
-
-        List<Integer> numbers = splittedNumbers.stream()
-                .map(String::trim)
-                .map(Integer::parseInt)
-                .collect(Collectors.toList());
-
-        return numbers;
-    }
-
     public static void printStatistics(LottoResult lottoResult, int purchaseAmount) {
-        Map<LottoRank, Integer> rankCounts = lottoResult.getRankCounts();
-        double profitRate = lottoResult.calculateProfitRate(purchaseAmount);
-
-        OutputView.printProfitRate(rankCounts, profitRate);
+        OutputView.printProfitRate(lottoResult, purchaseAmount);
     }
 }
